+++ cpp/include/autd3/modulation/transform.hpp
@@ -3,7 +3,7 @@
 // Created Date: 13/09/2023
 // Author: Shun Suzuki
 // -----
-// Last Modified: 21/09/2023
+// Last Modified: 26/09/2023
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -18,27 +18,30 @@
 namespace autd3::modulation {
 
 /**
- * @brief Modulation to cache the result of calculation
+ * @brief Modulation to transform the result of calculation
  */
-template <class M>
+template <class M, typename F>
 class Transform final : public internal::Modulation {
-  using transform_f = double (*)(uint32_t, double);
+  using transform_f = double (*)(const void*, uint32_t, double);
 
  public:
-  template <typename F>
-  Transform(M m, F& f) : _m(std::move(m)) {
-    _f_native = [f = std::move(f)](const uint32_t i, const double d) -> double { return f(static_cast<size_t>(i), d); };
+  Transform(M m, const F& f) : _m(std::move(m)), _f(f) {
+    _f_native = +[](const void* context, const uint32_t i, const double d) -> double {
+      return static_cast<const Transform*>(context)->_f(static_cast<size_t>(i), d);
+    };
   }
 
   AUTD3_IMPL_WITH_CACHE_MODULATION
 
   [[nodiscard]] internal::native_methods::ModulationPtr modulation_ptr() const override {
-    return internal::native_methods::AUTDModulationWithTransform(_m.modulation_ptr(), const_cast<void*>(static_cast<const void*>(&_f_native)));
+    return internal::native_methods::AUTDModulationWithTransform(_m.modulation_ptr(), const_cast<void*>(reinterpret_cast<const void*>(_f_native)),
+                                                                 const_cast<void*>(static_cast<const void*>(this)));
   }
 
  private:
   M _m;
-  std::function<double(uint32_t, double)> _f_native;
+  const F& _f;
+  transform_f _f_native;
 };
 
 }  // namespace autd3::modulation