+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix [#187](https://github.com/shinolab/autd3/issues/187): [C++] missing with_cache and with_transform for autd3::gain::holo::*
 - Fix [#188](https://github.com/shinolab/autd3/issues/188): [C++] autd3::gain::holo::LM::with_k_max type is invalid
 - Fix [#189](https://github.com/shinolab/autd3/issues/189): [C++] autd3::gain::holo::Greedy cause an SEHException
+- Fix [#190](https://github.com/shinolab/autd3/issues/190): [C++] Modulation::with_transform cause an SEH exception
 
 # 15.1.2